bugfix for chrono stage imports
